+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Change Log
 
+- csharp-mvc: support vsdx in diagram editor
 - csharp: support vsdx in diagram editor
 - nodejs: support vsdx in diagram editor
 - golang: new integration example